# docs/20-api/openapi.yaml
openapi: 3.1.0
info:
  title: ionowu Core API — POS Inti
  version: "0.1.0"
  description: >
    Kontrak API untuk Core Service (Golang) — Fase 0 + Fase 1.
    Sumber kebenaran untuk generate shared-types TypeScript (apps/web) dan
    handler Go (apps/core). Setiap perubahan di sini mendahului kode,
    bukan menyusul. Referensi: FDR.md §8, PRD-01-POS-INTI.md §8.
  contact:
    name: ionowu.com
servers:
  - url: https://api.sweet.ionowu.com/v1
    description: Produksi
  - url: http://localhost:8080/v1
    description: Lokal (Docker Compose)
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Daftarkan tenant baru + akun owner pertama
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [organization_name, outlet_name, owner_email, owner_password, owner_name]
              properties:
                organization_name: { type: string, minLength: 1, maxLength: 200 }
                outlet_name: { type: string, minLength: 1, maxLength: 200 }
                owner_email: { type: string, format: email }
                owner_password: { type: string, minLength: 10 }
                owner_name: { type: string, minLength: 1, maxLength: 200 }
                business_type:
                  type: string
                  maxLength: 40
                  description: >-
                    Slug jenis usaha, tersimpan di kolom tenants.business_type.
                    Opsional — kolomnya nullable dan pendaftaran tidak ditahan
                    kalau pemilik melewatinya. Daftar slug yang sah ada di
                    services/pos-engine/internal/businesstype/catalog.json
                    (berkas yang sama dibaca frontend), diturunkan dari
                    MARKET-SEGMENTS.md §3. Contoh: parfum_refill, toko_bahan_kue.
                  example: parfum_refill
      responses:
        "201":
          description: Tenant terdaftar
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400":
          description: business_type tidak ada di katalog (INVALID_BUSINESS_TYPE)
          content:
            application/json: { schema: { $ref: "#/components/schemas/Error" } }
        "409":
          description: Email sudah terdaftar
          content:
            application/json: { schema: { $ref: "#/components/schemas/Error" } }
  /auth/login:
    post:
      tags: [auth]
      summary: >-
        Login dengan email + password. Tenant TIDAK perlu disebut: idx_users_email
        UNIQUE global, jadi satu email hanya menunjuk satu user dan tenant-nya
        diambil dari baris user itu — bukan dari kiriman klien.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                tenant_id:
                  type: string
                  deprecated: true
                  description: >-
                    Masih diterima agar klien lama tidak pecah, tetapi diabaikan.
                    Tenant selalu diambil dari baris user hasil lookup email.
      responses:
        "200":
          description: Berhasil login
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "401":
          description: Kredensial salah
          content:
            application/json: { schema: { $ref: "#/components/schemas/Error" } }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Tukar refresh token dengan access token baru
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: Token baru
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "401":
          description: Refresh token tidak valid/dicabut
          content:
            application/json: { schema: { $ref: "#/components/schemas/Error" } }
  /auth/logout:
    post:
      tags: [auth]
      summary: Cabut refresh token aktif
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "204": { description: Berhasil logout }

  # ═══════════════════════════════════════════════════════════════
  # PRODUCTS — FR-10..15
  # ═══════════════════════════════════════════════════════════════
  /products:
    get:
      tags: [products]
      summary: Daftar produk (untuk manajemen katalog, bukan pencarian kasir)
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { name: category, in: query, schema: { type: string } }
        - { name: is_active, in: query, schema: { type: boolean } }
        - { name: q, in: query, description: "cari by nama/SKU", schema: { type: string } }
      responses:
        "200":
          description: Daftar produk
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Product" } }
                  page: { type: integer }
                  page_size: { type: integer }
                  total: { type: integer }
    post:
      tags: [products]
      summary: Buat produk baru (owner/manager saja)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "201":
          description: Produk dibuat
          content:
            application/json: { schema: { $ref: "#/components/schemas/Product" } }
        "403": { $ref: "#/components/responses/Forbidden" }
  /products/{productId}:
    parameters:
      - { name: productId, in: path, required: true, schema: { $ref: "#/components/schemas/ULID" } }
    patch:
      tags: [products]
      summary: Ubah produk, termasuk soft-delete (is_active=false, FR-15)
      description: >
        Owner & manager (RBAC-MODEL "Tambah/ubah produk"). Produk nonaktif membuat
        SEMUA variannya `is_active=false` di /sync/pull, jadi hilang dari kasir.
        Id salah dan id milik tenant lain sama-sama 404.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductPatch" }
      responses:
        "200":
          description: Produk diperbarui
          content:
            application/json: { schema: { $ref: "#/components/schemas/Product" } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: PRODUCT_NOT_FOUND }
        "422": { description: "VALIDATION_ERROR (nama kosong / > 200 huruf)" }
  /products/import:
    post:
      tags: [products]
      summary: Import produk massal via CSV (FR-14)
      description: |
        **Sinkron.** Sebelumnya kontrak ini menjanjikan `202` + `job_id` (asinkron), padahal
        implementasinya selalu sinkron dan mengembalikan `201`. Kontrak disamakan dengan
        kenyataan pada 2026-09-16.

        Kolom CSV (baris pertama header, dilewati):
        `ProductName, VariantName, SKU, Barcode, Price, CostPrice, Uom, UomPrecision, StockQuantity, ItemType, StockUom, StockFactor`

        `StockUom` + `StockFactor` opsional (ADR-0012): satuan STOK bila berbeda dari satuan
        jual — bibit dijual `ml`, stoknya `g`, faktor 1 (gram per ml, bawaan 1). Bila diisi,
        `StockQuantity` dibaca dalam satuan stok dan boleh sampai 3 desimal.

        Setiap baris diproses dalam SAVEPOINT sendiri: baris yang gagal dilewati dan
        **dilaporkan** di `dilewati`, baris lain tetap masuk. `Price` wajib diisi — harga
        kosong ditolak, bukan dianggap Rp 0.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Impor selesai
          content:
            application/json:
              schema:
                type: object
                required: [message, total_records, dilewati]
                properties:
                  message: { type: string, example: Impor selesai }
                  total_records:
                    type: integer
                    description: Jumlah baris yang BERHASIL masuk
                  dilewati:
                    type: array
                    description: Baris yang tidak diimpor beserta alasannya
                    items:
                      type: object
                      required: [baris, alasan]
                      properties:
                        baris:
                          type: integer
                          description: Nomor baris di berkas (header = 1)
                        alasan: { type: string, example: harga kosong atau bukan angka }
        "400":
          description: Berkas tidak ada, bukan CSV, atau hanya berisi header

  # ═══════════════════════════════════════════════════════════════
  # VARIANTS — FR-11..12
  # ═══════════════════════════════════════════════════════════════
...
  /variants/{variantId}:
    parameters:
      - { name: variantId, in: path, required: true, schema: { $ref: "#/components/schemas/ULID" } }
    patch:
      tags: [variants]
      summary: Ubah harga/HPP/barcode/batas stok varian
      description: >
        Owner & manager; `price` dan `cost_price` hanya owner (RBAC-MODEL
        "Ubah HPP & harga jual"). Kolom yang tidak dikirim tidak berubah.
        `sku`/`barcode` = "" mengosongkan (NULL). Angka dikirim sebagai string
        desimal; nilai yang tidak bisa dibaca DITOLAK 422, bukan diabaikan.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VariantPatch" }
      responses:
        "200": { description: Varian diupdate }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: VARIANT_NOT_FOUND }
        "409": { description: BARCODE_ALREADY_EXISTS }
        "422": { description: "VALIDATION_ERROR (angka tidak sah/negatif/terlalu banyak desimal)" }
...
  /sales:
    get:
      tags: [sales]
      summary: Riwayat transaksi satu outlet (histori kasir, bukan laporan BI)
      description: >
        Bawaan 30 hari terakhir, maksimal 200 baris, terbaru dulu. `refunded_total`
        ikut dihitung supaya layar riwayat bisa menandai nota yang sudah direfund.
        Data percobaan (`is_sandbox`) dikecualikan.
      parameters:
        - { name: outlet_id, in: query, required: true, schema: { $ref: "#/components/schemas/ULID" } }
        - { name: dari, in: query, schema: { type: string, format: date } }
        - { name: sampai, in: query, description: "inklusif", schema: { type: string, format: date } }
        - { name: cari, in: query, description: "cocokkan nomor nota", schema: { type: string, maxLength: 100 } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
      responses:
        "200":
          description: Daftar transaksi
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { $ref: "#/components/schemas/ULID" }
                        receipt_number: { type: string }
                        sold_at: { type: string, format: date-time }
                        grand_total: { type: string, format: decimal }
                        payment_status: { type: string, enum: [paid, void, pending] }
                        refunded_total: { type: string, format: decimal }
                        member_code: { type: string, nullable: true }
                        cashier_name: { type: string, nullable: true }
                        item_count: { type: integer }
        "422": { description: "VALIDATION_ERROR (outlet_id / tanggal)" }
    post:
      tags: [sales]
      summary: >
        Catat transaksi penjualan. IDEMPOTENT by client-generated ULID (FR-48) —
        request dengan id yang sama, berkali-kali, mengembalikan transaksi
        yang sama tanpa efek ganda. Ini endpoint yang dipanggil antrean sync
        klien setelah offline (FR-43).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SaleInput" }
      responses:
        "201":
          description: Transaksi baru tercatat
          content:
            application/json: { schema: { $ref: "#/components/schemas/Sale" } }
        "200":
          description: >
            Transaksi dengan id ini sudah pernah tercatat (idempotency hit) —
            mengembalikan record yang sudah ada, bukan error.
          content:
            application/json: { schema: { $ref: "#/components/schemas/Sale" } }
        "422":
          description: Payload tidak valid (mis. discount > discount_max_pct_cashier tanpa discount_approved_by)
          content:
            application/json: { schema: { $ref: "#/components/schemas/Error" } }
  /sales/{saleId}:
    get:
      tags: [sales]
      summary: Detail nota — barang, pembayaran, dan riwayat refund
      parameters:
        - { name: saleId, in: path, required: true, schema: { $ref: "#/components/schemas/ULID" } }
      responses:
        "200": { description: "Detail transaksi (sale, items, payments, refunds)" }
        "404": { description: TRANSACTION_NOT_FOUND }
  /sales/{saleId}/void:
    post:
      tags: [sales]
      summary: Batalkan transaksi salah input; stok kembali dan tercatat di ledger
      description: >
        HANYA untuk transaksi lunas, belum pernah direfund, dan shift-nya MASIH TERBUKA —
        Z-Report yang sudah dicetak tidak pernah berubah. Setelah shift ditutup, pakai refund.
        Owner & manager langsung; kasir wajib `approver_user_id` + `pin` manager.
      parameters:
        - { name: saleId, in: path, required: true, schema: { $ref: "#/components/schemas/ULID" } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 3, maxLength: 500 }
                approver_user_id: { $ref: "#/components/schemas/ULID" }
                pin: { type: string }
      responses:
        "200": { description: Transaksi dibatalkan }
        "403": { description: "FORBIDDEN_ROLE · MANAGER_PIN_REQUIRED" }
        "404": { description: TRANSACTION_NOT_FOUND }
        "422": { description: "TRANSACTION_ALREADY_VOIDED · SHIFT_CLOSED · VALIDATION_ERROR (sudah direfund / alasan kosong)" }
  /sales/{saleId}/refund:
    parameters:
      - { name: saleId, in: path, required: true, schema: { $ref: "#/components/schemas/ULID" } }
    post:
      tags: [sales]
      summary: Refund sebagian/penuh (FR-26). Ditolak kalau di luar refund_window_hours tenant tanpa approval.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refund_type, reason, amount]
              properties:
                refund_type:
                  type: string
                  enum: [full, partial]
                  description: Wajib. Handler menolak nilai lain dengan VALIDATION_ERROR.
                reason: { type: string, minLength: 1 }
                amount: { type: number, format: decimal, exclusiveMinimum: 0 }
                shift_id:
                  { $ref: "#/components/schemas/ULID" }
                restock:
                  type: boolean
                  default: false
                  description: Bila true, stok varian yang direfund dikembalikan.
                items:
                  type: array
                  description: Baris yang direfund. Dibutuhkan untuk restock per varian.
                  items:
                    type: object
                    required: [variant_id, quantity]
                    properties:
                      variant_id: { $ref: "#/components/schemas/ULID" }
                      quantity: { type: number, format: decimal, exclusiveMinimum: 0 }
                approver_user_id:
                  { $ref: "#/components/schemas/ULID" }
                pin:
                  type: string
                  description: >-
                    approver_user_id + pin WAJIB bila aktor berperan `cashier`
                    (RBAC-MODEL §"Void transaksi"). Owner/manager yang sedang
                    login boleh menyetujui sendiri; keduanya diabaikan.
      responses:
        "201":
          description: Refund tercatat
          content:
            application/json: { schema: { $ref: "#/components/schemas/Refund" } }
        "403":
          description: Di luar window refund, butuh approval manager
          content:
            application/json: { schema: { $ref: "#/components/schemas/Error" } }

  # ═══════════════════════════════════════════════════════════════
  # STOCK — FR-30..35
  # ═══════════════════════════════════════════════════════════════
  /stock/events:
    post:
      tags: [stock]
      summary: Satu mutasi stok — barang masuk (restock) atau rusak/hilang (waste)
      description: >
        Dipakai layar Stok. Owner, manager, dan gudang saja (RBAC-MODEL §Stok).
        `quantity` SELALU positif dan dalam SATUAN STOK (gram untuk bibit, ADR-0012);
        arah ditentukan `event_type`. Saldo `variants.stock_quantity` dan baris ledger
        ditulis dalam satu transaksi. `opname_adjust` tidak diterima di sini —
        koreksi opname hanya lahir dari `/stock/opname`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outlet_id, variant_id, event_type, quantity]
              properties:
                outlet_id: { $ref: "#/components/schemas/ULID" }
                variant_id: { $ref: "#/components/schemas/ULID" }
                event_type: { type: string, enum: [restock, waste] }
                quantity: { type: string, format: decimal, example: "250", description: "> 0, maks. 3 desimal, satuan stok" }
                note: { type: string, maxLength: 500 }
      responses:
        "201":
          description: Stok dicatat
          content:
            application/json:
              schema:
                type: object
                properties:
                  event_id: { $ref: "#/components/schemas/ULID" }
                  data:
                    type: object
                    properties:
                      stock_quantity: { type: string, format: decimal, description: "saldo baru" }
                      uom: { type: string, description: "satuan stok" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: "VARIANT_NOT_FOUND (tidak ada / tenant lain / tanpa stok) · OUTLET_NOT_FOUND" }
        "422": { description: "VALIDATION_ERROR (jumlah, event_type, id)" }
...
  /stock/levels:
    get:
      tags: [stock]
      summary: Level stok saat ini per outlet (dari materialized view, bisa negatif — lihat D-11)
      parameters:
        - { name: outlet_id, in: query, required: true, schema: { $ref: "#/components/schemas/ULID" } }
        - { name: variant_id, in: query, schema: { $ref: "#/components/schemas/ULID" } }
        - { name: only_negative, in: query, description: "FR-34: filter untuk daftar peringatan manager", schema: { type: boolean } }
      responses:
        "200":
          description: Level stok
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/StockLevel" } }
  /stock/opname:
    post:
      tags: [stock]
      summary: Opname — hasil timbang menjadi stok yang berlaku, selisihnya masuk ledger
      description: >
        Owner, manager, dan gudang saja. `counted_quantity` dalam SATUAN STOK.
        Jumlah menurut sistem TIDAK dikirim klien: ia dibaca di dalam transaksi
        sambil mengunci baris varian, jadi penjualan yang masuk bersamaan tidak
        hilang tertimpa. Selisih 0 tidak menghasilkan baris ledger.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outlet_id, items]
              properties:
                outlet_id: { $ref: "#/components/schemas/ULID" }
                items:
                  type: array
                  minItems: 1
                  maxItems: 500
                  items:
                    type: object
                    required: [variant_id, counted_quantity]
                    properties:
                      variant_id: { $ref: "#/components/schemas/ULID" }
                      counted_quantity: { type: string, format: decimal, example: "238.5" }
      responses:
        "201":
          description: Opname tersimpan
          content:
            application/json:
              schema:
                type: object
                properties:
                  opname_id: { $ref: "#/components/schemas/ULID" }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        variant_id: { $ref: "#/components/schemas/ULID" }
                        system: { type: string, format: decimal }
                        counted: { type: string, format: decimal }
                        variance: { type: string, format: decimal }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: "VARIANT_NOT_FOUND · OUTLET_NOT_FOUND" }
        "422": { description: "VALIDATION_ERROR (jumlah, barang ganda, id)" }
...
  /shifts/open:
    post:
      tags: [shifts]
      summary: >
        Buka shift kasir (input kas awal). `id` digenerate KLIEN (D-02) —
        shift wajib bisa dibuka saat perangkat offline, karena kasir yang
        tiba pagi hari tanpa koneksi harus tetap bisa mulai bekerja (FR-45).
        Saat offline, klien mencatat shift lokal lalu mengirimnya lewat
        /sync/push (field `shifts_open`), bukan lewat endpoint ini.
        Idempotent by id.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShiftOpenInput" }
      responses:
        "201":
          description: Shift terbuka
          content:
            application/json: { schema: { $ref: "#/components/schemas/Shift" } }
        "200":
          description: Shift dengan id ini sudah tercatat (idempotency hit)
          content:
            application/json: { schema: { $ref: "#/components/schemas/Shift" } }
        "409":
          description: Kasir masih punya shift terbuka lain
          content:
            application/json: { schema: { $ref: "#/components/schemas/Error" } }
  /shifts/{shiftId}/close:
    parameters:
      - { name: shiftId, in: path, required: true, schema: { $ref: "#/components/schemas/ULID" } }
    post:
      tags: [shifts]
      summary: Tutup shift — kirim kas fisik aktual, server hitung selisih (FR-51..52)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [closing_cash, occurred_at]
              properties:
                closing_cash: { type: number, format: decimal, minimum: 0 }
                occurred_at: { type: string, format: date-time }
      responses:
        "200":
          description: Shift tertutup dengan ringkasan
          content:
            application/json: { schema: { $ref: "#/components/schemas/ShiftSummary" } }
...
  /outlets:
    get:
      tags: [outlets]
      summary: Daftar outlet tenant
      responses:
        "200":
          description: Daftar outlet
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Outlet" } }
    post:
      tags: [outlets]
      summary: Buat outlet baru (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OutletInput" }
      responses:
        "201":
          description: Outlet dibuat
          content:
            application/json: { schema: { $ref: "#/components/schemas/Outlet" } }
        "403": { $ref: "#/components/responses/Forbidden" }
  /outlets/{id}:
    patch:
      tags: [outlets]
      summary: Ubah profil toko & struk
      description: >
        Owner: semua outlet. Manager: hanya outlet tempat ia ditugaskan, dan
        tidak boleh mengubah `is_active` (RBAC-MODEL §Matriks). Kolom yang
        tidak dikirim tidak berubah; string kosong mengosongkan kolom opsional.
        Id salah dan id milik tenant lain sama-sama dijawab 404.
      parameters:
        - { name: id, in: path, required: true, schema: { $ref: "#/components/schemas/ULID" } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
                address: { type: string, maxLength: 200 }
                phone: { type: string, maxLength: 50 }
                receipt_footer: { type: string, maxLength: 200, description: "Teks penutup struk; kosong = \"Terima kasih\"" }
                warranty_days: { type: integer, minimum: 0, maximum: 365, description: "Garansi (hari) yang dicetak di nota; 0 = tanpa garansi" }
                social_handle: { type: string, maxLength: 100, description: "Akun TikTok toko; \"@\" di depan dibuang" }
                bibit_percent: { type: integer, minimum: 0, maximum: 100, description: "Persen bibit racikan parfum (sisanya pelarut); 0 = tanpa racikan" }
                nota_web_url: { type: string, maxLength: 150, description: "https:// tanpa ?/#; \"\" = nota tanpa QR (ADR-0013)" }
                timezone: { type: string }
                business_day_start: { type: string, example: "00:00" }
                is_active: { type: boolean, description: "Owner saja" }
      responses:
        "200": { description: Outlet diupdate }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: OUTLET_NOT_FOUND }
        "422": { description: "VALIDATION_ERROR (nama kosong, teks terlalu panjang)" }

  # ═══════════════════════════════════════════════════════════════
  # NOTA PUBLIK — ADR-0013 (TANPA login)
  # ═══════════════════════════════════════════════════════════════
...
  /public/v1/nota/{tenantId}/{saleId}/member:
    parameters:
      - { name: tenantId, in: path, required: true, schema: { $ref: "#/components/schemas/ULID" } }
      - { name: saleId, in: path, required: true, schema: { $ref: "#/components/schemas/ULID" } }
    post:
      tags: [public]
      summary: Daftar member berbekal nota (satu pendaftaran per nota)
      description: >
        Tanpa login. Hanya untuk nota lunas, belum direfund, tanpa member, ≤ 30 hari.
        Rate limit 5/menit per nota dan 30/menit per tenant.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, follows_store_social]
              properties:
                phone: { type: string, example: "0812-3456-7890" }
                name: { type: string, maxLength: 200 }
                social_handle: { type: string, maxLength: 100 }
                follows_store_social: { type: boolean, description: "Wajib true (syarat member toko)" }
      responses:
        "201":
          description: Member terdaftar
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: object, properties: { member_code: { type: string, example: "M-8ER68X" } } }
        "404": { description: NOTA_NOT_FOUND }
        "409": { description: "NOTA_SIGNUP_CLOSED · MEMBER_ALREADY_EXISTS" }
        "422": { description: "VALIDATION_ERROR (nomor WA, belum follow, terlalu panjang)" }
        "429": { description: RATE_LIMITED }

  # ═══════════════════════════════════════════════════════════════
  # SYNC — FR-40..48 (endpoint paling kritis di seluruh API)
  # ═══════════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Forbidden:
      description: Role tidak punya izin (mis. kasir mencoba hapus produk — D-10)
      content:
        application/json: { schema: { $ref: "#/components/schemas/Error" } }
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: "VALIDATION_ERROR" }
        message: { type: string }
        details: { type: object, additionalProperties: true, nullable: true }
    AuthSession:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        expires_in: { type: integer, example: 900, description: "detik, default 15 menit" }
        user:
          type: object
          properties:
            id: { $ref: "#/components/schemas/ULID" }
            tenant_id: { $ref: "#/components/schemas/ULID" }
            email: { type: string, format: email }
            role: { type: string, enum: [owner, manager, cashier] }
    Product:
      type: object
      properties:
        id: { $ref: "#/components/schemas/ULID" }
        sku: { type: string, nullable: true }
        name: { type: string }
        category: { type: string, nullable: true }
        unit: { type: string, example: "pcs" }
        image_url: { type: string, format: uri, nullable: true }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ProductInput:
      type: object
      required: [name, unit]
      properties:
        sku: { type: string, nullable: true }
        name: { type: string, minLength: 1, maxLength: 200 }
        category: { type: string, nullable: true }
        unit: { type: string, default: "pcs" }
        image_url: { type: string, format: uri, nullable: true }
...